GSA-99 Mark survey version dirty after editing
Changes made to a published survey's questions are not reflected in the
reporting interface until after the regeneration task is run.  Until the task
runs, any existing reports may have incorrectly named questions and/or
invalid answers due to naming changes.

Running the daily report generation task (reporting:daily) will update the
report fields that were edited and mark the survey version as clean.

A notice of the dirty report status on the reporting landing page until the
daily task has been run and the survey version marked as clean.