Move export to mongo to after processing
As an after_save callback, the mongo processing was happening before the
display field values had been populated.  This was making the mongo
export contain only nil values for the answers.  Moving the export to be
triggered at the end of #process_me ensures that the mongo export will
occur after the display field values have been populated.