Fix bug: SurveyVersion import choice questions
SurveyVersion import was not correctly importing the value for the
ChoiceQuestion answer_placement field.  This resulted in imported
surveys not having the correct classes applied to the generated HTML for
ChoiceQuestions where the answer_placement field should have been true.

The fix was to add the answer_placement field to the importer so it
would be correctly set on the imported survey version's choice
questions.

Fixes Intervals ticket #1985.