Add rake task reporting:export_in_range
To avoid re-exporting all survey responses to mongo and elasticsearch,
the export_in_range task takes a start and end date and exports only the
survey responses with a created_at date within that range.

rake reporting:export_in_range["2015-08-01", "2015-08-31"]

An error will be raised if either the start date or end date are not
supplied or not in the correct format for parsing.