Fix matrixquestion creation issue
Creating a MatrixQuestion deletes all ChoiceAnswers for ChoiceQuestions
which do not belong to a MatrixQuestion.  MatrixQuestion has an
after_validation callback that is used to remove ChoiceAnswers from the
sub ChoiceQuestions before adding the new ChoiceAnswers.  However this
is run when the action is not an update action.  This causes the code
`self.choice_questions.includes(:choice_answers)` to generate the SQL:
`SELECT * FROM choice_questions WHERE
choice_questions.matrix_question_id = NULL` which returns all
ChoiceQuestions which are not part of a MatrixQuestion and causes the
callback function to delete all of their ChoiceAnswers.

The callback is now changed from an after_validation to a before_update
in order to only be called when the MatrixQuestion has already been
persisted to the database and the id is not NULL.