Remove null constraint on the old name fields
Do not remove the f_name and l_name columns.  Instead, remove the null
constraint in case we ever need to roll back.

Update the user factory to create a full name instead of using f_name
and l_name

update the feature helper to find the link based on the hhs_id instead
of by user name