Use jQuery.on() to attach callbacks dynamically
The page updates the question list using javascript.  Updating the page
after the first page load causes the onClick callbacks to no longer be
bound to the delete links as they were reloaded.

Using the .on() method to dynamically bind to all deleteLinks on the
page no matter when they were added causes the binding to properly be
added when the page is updated.