Catch the Faraday connection refused exception
When booting the application, if we cannot establish a connection to the
ElasticSearch server Faraday will throw a connection exception.  This
will halt the rails boot process and dump a stacktrace to the console
that is far from helpful.

Catch the Faraday connection refused exception and print a useful
message to the console letting the user know that we cannot connect to
ElasticSearch before exiting gracefully.